requirements.txt
# Core
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM
litellm>=1.0.0
google-genai>=1.0.0

# Memory
mem0ai>=0.1.0
qdrant-client>=1.7.0

# Templates
jinja2>=3.0.0
